Fix other extension code + tests
