+++ BrowserKit/Sources/Common/Extensions/StringExtension.swift
@@ -39,4 +39,17 @@ public extension String {
         let filteredWords = words.filter { !$0.isEmpty }
         return Int32(clamping: filteredWords.count)
     }
+
+    /// Encode HTMLStrings
+    /// Also used for Strings which are not sanitized for displaying
+    /// - Returns: Encoded String
+    var htmlEntityEncodedString: String {
+        return self
+            .replacingOccurrences(of: "&", with: "&amp;", options: .literal)
+            .replacingOccurrences(of: "\"", with: "&quot;", options: .literal)
+            .replacingOccurrences(of: "'", with: "&#39;", options: .literal)
+            .replacingOccurrences(of: "<", with: "&lt;", options: .literal)
+            .replacingOccurrences(of: ">", with: "&gt;", options: .literal)
+            .replacingOccurrences(of: "`", with: "&lsquo;", options: .literal)
+    }
 }
