+++ BrowserKit/Sources/Common/Extensions/URLExtension.swift
@@ -308,9 +308,67 @@ extension URL {
         return nil
     }
 
+    // MARK: Reader mode
+
     public var isReaderModeURL: Bool {
         let scheme = self.scheme, host = self.host, path = self.path
         return scheme == "http" && host == "localhost" && path == "/reader-mode/page"
+    }
+
+    public var decodeReaderModeURL: URL? {
+        if self.isReaderModeURL || self.isSyncedReaderModeURL {
+            if let components = URLComponents(url: self, resolvingAgainstBaseURL: false),
+               let queryItems = components.queryItems {
+                if let queryItem = queryItems.first(where: { $0.name == "url" }),
+                   let value = queryItem.value {
+                    return URL(string: value)?.safeEncodedUrl
+                }
+            }
+        }
+        return nil
+    }
+
+    public var isSyncedReaderModeURL: Bool {
+        return absoluteString.hasPrefix("about:reader?url=")
+    }
+
+    public var safeEncodedUrl: URL? {
+        var components = URLComponents(url: self, resolvingAgainstBaseURL: false)
+
+        // HTML-encode scheme, host, and path
+        guard let host = components?.host?.htmlEntityEncodedString,
+              let scheme = components?.scheme?.htmlEntityEncodedString,
+              let path = components?.path.htmlEntityEncodedString else {
+            return nil
+        }
+
+        components?.path = path
+        components?.scheme = scheme
+        components?.host = host
+
+        // sanitize query items
+        if let queryItems = components?.queryItems {
+            var safeQueryItems: [URLQueryItem] = []
+
+            for item in queryItems {
+                // percent-encoded characters
+                guard let decodedValue = item.value?.removingPercentEncoding else {
+                    return nil
+                }
+
+                // HTML special characters
+                let htmlEncodedValue = decodedValue.htmlEntityEncodedString
+
+                // New query item with the HTML-encoded value
+                let safeItem = URLQueryItem(name: item.name, value: htmlEncodedValue)
+                safeQueryItems.append(safeItem)
+            }
+
+            // Replace the original query items with the "safe" ones
+            components?.queryItems = safeQueryItems
+        }
+
+        return components?.url
     }
 }
