+++ firefox-ios/Client/Configuration/version.xcconfig
@@ -1,1 +1,1 @@
-APP_VERSION = 142.1
+APP_VERSION = 143.0
